UserProfil mitgeben Filter dynamisch float to Double

+++ OlympGuide.Application/Features/SportField/SportFieldDTO.cs
@@ -1,4 +1,4 @@
 ﻿namespace OlympGuide.Application.Features.SportField
 {
-    public record SportFieldDto(Guid Id, string Name, string Description, float Longitude, float Latitude, string Address);
+    public record SportFieldDto(Guid Id, string Name, string Description, double Longitude, double Latitude, string Address);
 }
